perpare to move to scenario based testing
